Move Subversion arg unit tests from functional folder to unit test folder
